+++ doc/release_notes.md
@@ -115,3 +115,6 @@ OpenPose Library - Release Notes
 
 ## Current version (future OpenPose 1.1.1)
 1. Main improvements:
+    1. COCO JSON file outputs 0 as score for non-detected keypoints.
+    2. Added example for OpenPose for user asynchronous output and cleaned all `tutorial_wrapper/` examples.
+    3. Added `-1` option for `net_resolution` in order to auto-select the best possible aspect ratio given the user input.